Added G4IOG-EU 10m grabber.
In addition to the usual grabber facilities, G4IOG has added a Raspberry Pi 3
and an RTL SDR dongle to perform split grabbing. E.G. two bands within 10m
on the same dongle.

This system works from a separate vertical antenna and is also characterized
by having a 10 minute grab time as opposed to the others that have a 20 minute cycle.